+++ test/com/amplitude/api/IdentifyTest.java
@@ -329,4 +329,42 @@ public class IdentifyTest extends BaseTest {
         expected.put(Constants.AMP_OP_SET_ONCE, new JSONObject().put(property, value1));
         assertTrue(Utils.compareJSONObjects(expected, identify.userPropertiesOperations));
     }
+
+    @Test
+    public void testGetUserPropertyOperations() throws JSONException {
+        String property1 = "string value";
+        String value1 = "testValue";
+
+        String property2 = "double value";
+        double value2 = 0.123;
+
+        String property3 = "boolean value";
+        boolean value3 = true;
+
+        String property4 = "json value";
+
+        String property5 = "array value";
+        JSONArray value5 = new JSONArray();
+        value5.put(15);
+        value5.put(25);
+
+        String property6 = "int value";
+        int value6 = 100;
+
+        Identify identify = new Identify().setOnce(property1, value1).add(property2, value2);
+        identify.set(property3, value3).unset(property4).append(property5, value5);
+        identify.prepend(property6, value6);
+
+        // identify should ignore this since duplicate key
+        identify.set(property4, value3);
+
+        JSONObject expected = new JSONObject();
+        expected.put(Constants.AMP_OP_SET_ONCE, new JSONObject().put(property1, value1));
+        expected.put(Constants.AMP_OP_ADD, new JSONObject().put(property2, value2));
+        expected.put(Constants.AMP_OP_SET, new JSONObject().put(property3, value3));
+        expected.put(Constants.AMP_OP_UNSET, new JSONObject().put(property4, "-"));
+        expected.put(Constants.AMP_OP_APPEND, new JSONObject().put(property5, value5));
+        expected.put(Constants.AMP_OP_PREPEND, new JSONObject().put(property6, value6));
+        assertTrue(Utils.compareJSONObjects(expected, identify.getUserPropertiesOperations()));
+    }
 }